BT-APP: GATT: Fix characteristic, descriptor size type
In write characteristic and descriptor requests the length of data
is stored in uint8_t type, this is causing overflow when length
of data is more then 255. Fix this by changing len type to size_t

+++ bt-app/gatt/lib/src/GattLibService.cpp
@@ -2794,7 +2794,7 @@ void GattLibService::HandleGattsRequestWriteCharacteristicEvent(
                                                        GattsRequestWriteCharacteristicEvent *event)
 {
   if (!sGattService) return;
-  uint8_t len = event->value->size();
+  size_t len = event->value->size();
   uint8_t *p_value = new uint8_t[len+1];
   if (len == 0) {
     ALOGE(LOGTAG "HandleGattsRequestWriteCharacteristicEvent () - Data is NULL");
@@ -2812,7 +2812,7 @@ void GattLibService::HandleGattsRequestWriteDescriptorEvent(
                                                       GattsRequestWriteDescriptorEvent *event)
 {
   if (!sGattService) return;
-  uint8_t len = event->value->size();
+  size_t len = event->value->size();
   uint8_t *p_value = new uint8_t[len+1];
   if (len == 0) {
     ALOGE(LOGTAG "HandleGattsRequestWriteDescriptorEvent () - Data is NULL");
